+++ src/main/java/edu/escuelaing/arsw/persistence/UserPersistence.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 
 
 @Repository
-public interface UserPersistence extends CrudRepository<User, Integer>{
+public interface UserPersistence extends CrudRepository<User, Long>{
     public User findByEmail(String email);
 }
